openmp 2.0 compat

+++ src/Harvester.cpp
@@ -70,7 +70,7 @@ void Harvester::handle(std::shared_ptr<const Challenge> value)
 	std::vector<std::vector<uint128_t>> scores(plots.size());
 
 #pragma omp parallel for num_threads(num_threads)
-	for(size_t i = 0; i < plots.size(); ++i)
+	for(int i = 0; i < int(plots.size()); ++i)
 	{
 		const auto& prover = plots[i];
 		try {
@@ -196,7 +196,7 @@ void Harvester::reload()
 	std::vector<std::pair<std::shared_ptr<vnx::File>, std::shared_ptr<chiapos::DiskProver>>> plots;
 
 #pragma omp parallel for num_threads(num_threads)
-	for(size_t i = 0; i < all_dirs.size(); ++i)
+	for(int i = 0; i < int(all_dirs.size()); ++i)
 	{
 		vnx::Directory dir(all_dirs[i]);
 		try {
