+++ src/Node_update.cpp
@@ -139,7 +139,7 @@ void Node::update()
 	}
 
 #pragma omp parallel for if(!is_synced)
-	for(size_t i = 0; i < to_verify.size(); ++i)
+	for(int i = 0; i < int(to_verify.size()); ++i)
 	{
 		const auto& fork = to_verify[i];
 		const auto& block = fork->block;
@@ -490,7 +490,7 @@ bool Node::make_block(std::shared_ptr<const BlockHeader> prev, std::shared_ptr<c
 	context->height = block->height;
 
 #pragma omp parallel for
-	for(size_t i = 0; i < tx_list.size(); ++i)
+	for(int i = 0; i < int(tx_list.size()); ++i)
 	{
 		auto& entry = tx_list[i];
 		auto& tx = entry.tx;
